fix: Conda venv is not used
Conda doesn't follow PEP 405.
